requirements.txt
# Chargebee API
chargebee>=3.0.0

# Google Sheets
gspread>=6.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# Environment
python-dotenv>=1.0.0
